a2ui-demo/samples/agent/adk/simple_chat/pyproject.toml
[project]
name = "a2ui-simple-chat"
version = "0.1.0"
description = "Simple A2UI chat agent - a minimal chat application using Gemini."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "a2ui-agent",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
